requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
numpy==1.24.3
redis==5.0.1
PyYAML==6.0.1
clip-client==0.1.0
